250904新增會員管理與修正訂單頁面

+++ src/main/java/com/petpick/petpick/controller/shop/OrderController.java
@@ -27,9 +27,10 @@ public class OrderController {
 
     /** 訂單總覽（某使用者） */
     @GetMapping("/user/{userId}")
-    public List<OrderSummaryDTO> list(@PathVariable Integer userId) {
+    public List<OrderSummaryDTO> list(@PathVariable Long userId) {
         return orderQueryService.listByUser(userId);
     }
+
 
     /** 訂單抬頭＋可能含明細 */
     @GetMapping("/{orderId}")
